вынес app bar в виджеты
